rewrite lammps reader
